+++ crypto/decrepit/cast/cast.c
@@ -64,7 +64,7 @@ OPENSSL_MSVC_PRAGMA(warning(push, 3))
 OPENSSL_MSVC_PRAGMA(warning(pop))
 #endif
 
-#include "../../crypto/internal.h"
+#include "../../internal.h"
 #include "internal.h"
 #include "../macros.h"
