+++ crypto/decrepit/ripemd/ripemd_test.cc
@@ -21,8 +21,8 @@
 
 #include <gtest/gtest.h>
 
-#include "../../crypto/internal.h"
-#include "../../crypto/test/test_util.h"
+#include "../../internal.h"
+#include "../../test/test_util.h"
 
 
 struct RIPEMDTestCase {
